Fix firmware build for ESP-IDF v5.5
First real on-device compile (prior verification was sim + header checks
only). Three fixes for a clean `idf.py build`:

- idf_component.yml: pin esp_websocket_client to ~1.1.0; 1.2.x calls
  esp_transport_ws_get_redir_uri, absent in IDF 5.5's tcp_transport
- now_playing_screen.c: cast int32_t to int in fmt_time (int32_t is `long`
  on xtensa -> %d mismatch under -Werror=format)
- main/CMakeLists.txt: define LV_LVGL_H_INCLUDE_SIMPLE so generated
  ui/mdi_*.c fonts include "lvgl.h" instead of unresolvable "lvgl/lvgl.h"

Result: Project build complete, ytmusic_companion.bin generated, 0 warnings.

+++ ui/now_playing_screen.c
@@ -41,7 +41,8 @@ static void emit(const char *cmd, int arg) { if (s_emit) s_emit(cmd, arg); }
 static void fmt_time(char *buf, size_t n, int32_t sec)
 {
     if (sec < 0) sec = 0;
-    snprintf(buf, n, "%d:%02d", sec / 60, sec % 60);
+    // int32_t is `long` on the xtensa toolchain; cast so %d matches under -Werror=format.
+    snprintf(buf, n, "%d:%02d", (int)(sec / 60), (int)(sec % 60));
 }
 
 // ---- events ----
